+++ RemsLogic.Services/IComplianceService.cs
@@ -14,6 +14,6 @@ namespace RemsLogic.Services
         void AddEocsToPrescriberProfile(long profileId, long drugId);
         void RemoveEocsFromPrescriberProfile(long profileId, long drugId);
 
-        Dictionary<Drug, List<PrescriberEoc>> GetEocsByPrescriberProfile(int profileId);
+        Dictionary<Drug, List<PrescriberEoc>> GetEocsByPrescriberProfile(long profileId);
     }
 }
